+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/data_source/biometrics_disabled_local_data_source.dart
@@ -1,0 +1,11 @@
+import 'package:widget_toolkit_biometrics/widget_toolkit_biometrics.dart';
+
+/// A data source implementation used by default, when the biometrics
+/// authentication feature is not required.
+class BiometricsDisabledLocalDataSource implements BiometricsLocalDataSource {
+  @override
+  Future<bool> areBiometricsEnabled() async => false;
+
+  @override
+  Future<void> setBiometricsEnabled(bool enable) async => Future.value();
+}