+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/ui_components/pin_code_keyboard.dart
@@ -36,15 +36,17 @@ import 'pin_code_component.dart';
 /// will be stored in the device secure storage, the biometrics authentication
 /// will be automatically triggered and the biometrics icon will be displayed
 /// on the bottom right. When you press it every time it will trigger the
-/// biometric authentication.
+/// biometric authentication. Note: If [biometricsLocalDataSource] parameter is
+/// not provided to `PinCodeKeyboard` the biometrics authentication feature
+/// cannot be used.
 ///
 /// For more information on how the widgets work check the functional specification
 /// section in the README.md file.
 class PinCodeKeyboard extends StatelessWidget {
   const PinCodeKeyboard({
     required this.pinCodeService,
-    required this.biometricsLocalDataSource,
     required this.translateError,
+    this.biometricsLocalDataSource,
     this.mapBiometricMessageToString,
     this.isAuthenticatedWithBiometrics,
     this.isPinCodeVerified,
@@ -67,7 +69,9 @@ class PinCodeKeyboard extends StatelessWidget {
   final PinCodeService pinCodeService;
 
   /// Provides a contract to be implemented for the biometrics related methods.
-  final BiometricsLocalDataSource biometricsLocalDataSource;
+  /// If this parameter is not provided the biometrics authentication is disabled
+  /// for the package
+  final BiometricsLocalDataSource? biometricsLocalDataSource;
 
   /// Called when a user is authenticated with biometrics successfully
   final void Function(bool)? isAuthenticatedWithBiometrics;
@@ -117,6 +121,7 @@ class PinCodeKeyboard extends StatelessWidget {
                       pinLength: snapshot.data!,
                       translateError: translateError,
                       mapBiometricMessageToString: mapBiometricMessageToString,
+                      biometricsLocalDataSource: biometricsLocalDataSource,
                       isAuthenticatedWithBiometrics:
                           isAuthenticatedWithBiometrics,
                       isPinCodeVerified: isPinCodeVerified,