+++ packages/widget_toolkit_pin/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [0.0.2-dev.1]
 * Update `widget_toolkit` and `widget_toolkit_biometrics` dependencies
+* Update parameter `biometricsLocalDataSource` in `PinCodeKeyboard` to be optional
+* Add option to disable biometrics authentication, when not required
 
 ## [0.0.1-dev2]
 * Requires Dart `3.0`